+++ src/app/Security/Credentials/StoredCredentials.tsx
@@ -13,6 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+import { BreadcrumbPage } from '@app/BreadcrumbPage/BreadcrumbPage';
 import { DeleteWarningModal } from '@app/Modal/DeleteWarningModal';
 import { DeleteOrDisableWarningType } from '@app/Modal/types';
 import { JmxAuthDescription } from '@app/Shared/Components/JmxAuthDescription';
@@ -57,7 +58,6 @@ import _ from 'lodash';
 import * as React from 'react';
 import { concatMap, forkJoin, map } from 'rxjs';
 import { CreateCredentialModal } from './CreateCredentialModal';
-import { BreadcrumbPage } from '@app/BreadcrumbPage/BreadcrumbPage';
 
 export const includesCredential = (credentials: MatchedCredential[], credential: MatchedCredential): boolean => {
   return credentials.some((cred) => cred.id === credential.id);
